test: Add Plot.js to project, basic setup for Plot.plot(). Attention: code not running at the moment, this is intentional
The Plot library was added, which simplifies plotting a choropleth immensely.
Some preparation for the plotting was done.
However the data that has to be passed in has to be adjusted accordingly.
A dataset is needed that combines both geometrical data and joins at the fips identifier, which then leads to all counties having
a bachelorOrHigher percentage score, which then can be plotted.
The bachelorOrHigher will be the fill, the geometrical data obviously will create the US map with all of its
states and counties.